+++ src/screens/Detail/Segments/VoteResults.js
@@ -34,6 +34,7 @@ const VoteResults = props => {
               label !== "__typename" ? { value, label } : false
           ).filter(e => e)}
           colorScale={["#15C063", "#EC3E31", "#2C82E4"]}
+          label="Abstimmungen"
         />
       );
     }
@@ -56,6 +57,7 @@ const VoteResults = props => {
               label !== "__typename" ? { value, label } : false
           ).filter(e => e)}
           colorScale={["#99C93E", "#D43194", "#4CB0D8", "#B1B3B4"]}
+          label="Abgeordnete"
         />
       );
     }